Move matrix operations to separate file.
